% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_commands.R
\name{get_argument_description}
\alias{get_argument_description}
\title{Get Argument Description for Template}
\usage{
get_argument_description(template_name)
}
\arguments{
\item{template_name}{Name of the template}
}
\value{
Character string with argument descriptions
}
\description{
Returns a formatted description of arguments for each template
}
